+++ skfem/assembly/basis/facet_basis.py
@@ -3,6 +3,7 @@ from typing import Optional
 import numpy as np
 from numpy import ndarray
 
+from skfem.element import DiscreteField
 from skfem.quadrature import get_quadrature
 from .basis import Basis
 
@@ -11,45 +12,11 @@ class FacetBasis(Basis):
     """Global basis functions evaluated at integration points on the element
     boundaries.
 
-    Attributes
-    ----------
-    phi : ndarray
-        Global basis functions at global quadrature points.
-    dphi : ndarray
-        Global basis function derivatives at global quadrature points.
-    X : ndarray
-        Local quadrature points (Ndim x Nqp).
-    W : ndarray
-        Local quadrature weights (Nqp).
-    nf : int
-        Number of facets.
-    dx : ndarray
-        Used in computing global integrals elementwise (Nelems x Nqp).
-        For example, np.sum(u**2*dx, axis=1) where u is also
-        a numpy array of size Nelems x Nqp.
-    find : ndarray
-        A list of facet indices.
-    tind : ndarray
-        A list of triangle indices.
-    normals : ndarray
-    mapping : skfem.mapping.Mapping
-    elem : skfem.element.Element
-    Nbfun : int
-        The number of basis functions.
-    intorder : int
-        The integration order.
-    dim : int
-        The problem dimension.
-    nt : int
-        Number of triangles.
-    mesh : skfem.mesh.Mesh
-    refdom : string
-    brefdom : string
-
     Examples
     --------
-    FacetBasis object is a combination of Mesh, Element,
-    and Mapping:
+    :class:`~skfem.assembly.FacetBasis` object is a combination of
+    :class:`~skfem.mesh.Mesh`, :class:`~skfem.element.Element`, and
+    :class:`~skfem.mapping.Mapping`:
 
     >>> from skfem import *
     >>> from skfem.models.poisson import mass
@@ -102,10 +69,10 @@ class FacetBasis(Basis):
         # facets where the basis is evaluated
         if facets is None:
             if side is None:
-                self.find = np.nonzero(self.mesh.f2t[1, :] == -1)[0]
+                self.find = np.nonzero(self.mesh.f2t[1] == -1)[0]
                 self.tind = self.mesh.f2t[0, self.find]
             elif side == 0 or side == 1:
-                self.find = np.nonzero(self.mesh.f2t[1, :] != -1)[0]
+                self.find = np.nonzero(self.mesh.f2t[1] != -1)[0]
                 self.tind = self.mesh.f2t[side, self.find]
             else:
                 raise Exception("Parameter 'side' must be either 0 or 1. "
@@ -131,13 +98,15 @@ class FacetBasis(Basis):
                                                 self.find,
                                                 self.mesh.t2f)
 
+        self.normals = DiscreteField(self.normals)
+
         self.nelems = len(self.find)
 
         self.basis = [self.elem.gbasis(self.mapping, Y, j, self.tind)
                       for j in range(self.Nbfun)]
 
-        self.dx = np.abs(self.mapping.detDG(self.X, find=self.find)) *\
-            np.tile(self.W, (self.nelems, 1))
+        self.dx = (np.abs(self.mapping.detDG(self.X, find=self.find))
+                   * np.tile(self.W, (self.nelems, 1)))
 
         self.element_dofs = self.element_dofs[:, self.tind]
 
@@ -148,11 +117,9 @@ class FacetBasis(Basis):
                 'n': self.normals}
 
     def global_coordinates(self) -> ndarray:
-        return self.mapping.G(self.X, find=self.find)
+        return DiscreteField(self.mapping.G(self.X, find=self.find))
 
     def mesh_parameters(self) -> ndarray:
-        if self.mesh.dim() == 1:
-            return np.array([0.0])
-        else:
-            return (np.abs(self.mapping.detDG(self.X, self.find)) **
-                    (1.0 / (self.mesh.dim() - 1)))
+        return DiscreteField((np.abs(self.mapping.detDG(self.X, self.find))
+                              ** (1. / (self.mesh.dim() - 1.)))\
+                             if self.mesh.dim() != 1 else np.array([0.]))