+++ skfem/assembly/form/__init__.py
@@ -1,4 +1,4 @@
 from .form import Form
-from .bilinear_form import bilinear_form
-from .linear_form import linear_form
-from .functional import functional
+from .bilinear_form import bilinear_form, BilinearForm
+from .linear_form import linear_form, LinearForm
+from .functional import functional, Functional