+++ skfem/assembly/form/bilinear_form.py
@@ -1,97 +1,70 @@
-from typing import Any, Callable, Optional
+from typing import Any, Callable, Optional, Dict
 
 import numpy as np
 from numpy import ndarray
-from scipy.sparse import coo_matrix, csr_matrix
 
-from .form import Form, BasisTuple
-from .form_parameters import FormParameters
+from .form import Form, FormDict
 from ..basis import Basis
+from ...element import DiscreteField
 
 
 class BilinearForm(Form):
 
-    def kernel(self,
-               A: ndarray,
-               ix: ndarray,
-               ubasis: BasisTuple,
-               vbasis: BasisTuple,
-               w: FormParameters,
-               dx: ndarray) -> None:
-        for k in range(ix.shape[0]):
-            i, j = ix[k]
-            A[i, j] = np.sum(self.form(*ubasis[j], *vbasis[i], w) * dx,
-                             axis=1)
-
     def assemble(self,
-                 ubasis: Basis,
-                 vbasis: Optional[Basis] = None,
-                 w: Optional[Any] = (None, None, None),
-                 nthreads: Optional[int] = 1) -> csr_matrix:
-        """return sparse CSR matrix discretizing form
+                 u: Basis,
+                 v: Optional[Basis] = None,
+                 w: Dict[str, DiscreteField] = {}) -> Any:
 
-        :param w: A tuple of ndarrays. In the form definition:
+        if v is None:
+            v = u
+        elif u.intorder != v.intorder:
+            raise ValueError("Quadrature mismatch: trial and test functions "
+                             "should have same number of integration points.")
 
-          * :code:`w[0]` is accessible as :code:`w.w`,
-
-          * :code:`w[1]` is accessible as :code:`w.dw`, and
-
-          * :code:`w[2]` is accessible as :code:`w.ddw`.
-
-        The output of :meth:`~skfem.assembly.Basis.interpolate`
-        can be passed directly to this parameter.
-
-        """
-
-        import threading
-        from itertools import product
-
-        if vbasis is None:
-            vbasis = ubasis
-        else:
-            assert ubasis.intorder == vbasis.intorder, "Quadrature mismatch"
-
-        nt = ubasis.nelems
-        dx = ubasis.dx
-        w = self.parameters(w, ubasis)
+        nt = u.nelems
+        dx = u.dx
+        w = FormDict({**u.default_parameters(), **self.dictify(w)})
 
         # initialize COO data structures
-        data = np.zeros((vbasis.Nbfun, ubasis.Nbfun, nt))
-        rows = np.zeros(ubasis.Nbfun * vbasis.Nbfun * nt)
-        cols = np.zeros(ubasis.Nbfun * vbasis.Nbfun * nt)
+        sz = u.Nbfun * v.Nbfun * nt
+        data = np.zeros(sz)
+        rows = np.zeros(sz)
+        cols = np.zeros(sz)
 
-        # create sparse matrix indexing
-        for j in range(ubasis.Nbfun):
-            for i in range(vbasis.Nbfun):
-                # find correct location in data,rows,cols
-                ixs = slice(nt * (vbasis.Nbfun * j + i),
-                            nt * (vbasis.Nbfun * j + i + 1))
-                rows[ixs] = vbasis.element_dofs[i, :]
-                cols[ixs] = ubasis.element_dofs[j, :]
+        # loop over the indices of local stiffness matrix
+        for j in range(u.Nbfun):
+            for i in range(v.Nbfun):
+                ixs = slice(nt * (v.Nbfun * j + i),
+                            nt * (v.Nbfun * j + i + 1))
+                rows[ixs] = v.element_dofs[i]
+                cols[ixs] = u.element_dofs[j]
+                data[ixs] = self._kernel(u.basis[j], v.basis[i], w, dx)
 
-        # create indices for linear loop over local stiffness matrix
-        ixs = [i for j, i in product(range(ubasis.Nbfun), range(vbasis.Nbfun))]
-        jxs = [j for j, i in product(range(ubasis.Nbfun), range(vbasis.Nbfun))]
-        indices = np.array([ixs, jxs]).T
+        # TODO: allow user to change, e.g. cuda or petsc
+        return self._assemble_scipy_matrix(data, rows, cols, (v.N, u.N))
 
-        # split local stiffness matrix elements to threads
-        threads = [threading.Thread(
-            target=self.kernel,
-            args=(data, ij, ubasis.basis, vbasis.basis, w, dx))
-                   for ij in np.array_split(indices, nthreads, axis=0)]
-
-        # start threads and wait for finishing
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
-
-        K = coo_matrix((np.transpose(data, (1, 0, 2)).flatten('C'),
-                        (rows, cols)),
-                       shape=(vbasis.N, ubasis.N))
-        K.eliminate_zeros()
-        return K.tocsr()
+    def _kernel(self, u, v, w, dx):
+        return np.sum(self.form(u, v, w) * dx, axis=1)
 
 
 def bilinear_form(form: Callable) -> BilinearForm:
-    return BilinearForm(form)
+
+    # for backwards compatibility
+    from .form_parameters import FormParameters
+
+    class ClassicBilinearForm(BilinearForm):
+
+        def _kernel(self, u, v, w, dx):
+            W = {k: w[k].f for k in w}
+            if 'w' in w:
+                W['dw'] = w['w'].df
+            if u.ddf is not None:
+                return np.sum(self.form(u=u.f, du=u.df, ddu=u.ddf,
+                                        v=v.f, dv=v.df, ddv=v.ddf,
+                                        w=FormParameters(**W)) * dx, axis=1)
+            else:
+                return np.sum(self.form(u=u.f, du=u.df,
+                                        v=v.f, dv=v.df,
+                                        w=FormParameters(**W)) * dx, axis=1)
+
+    return ClassicBilinearForm(form)