+++ skfem/assembly/form/linear_form.py
@@ -1,65 +1,61 @@
-from typing import Any, Callable, Optional
+from typing import Any, Callable, Optional, Dict
 
 import numpy as np
 from numpy import ndarray
-from scipy.sparse import coo_matrix
 
-from .form import Form, BasisTuple
-from .form_parameters import FormParameters
+from .form import Form, FormDict
 from ..basis import Basis
+from ...element import DiscreteField
 
 
 class LinearForm(Form):
 
-    def kernel(self,
-               b: ndarray,
-               ix: ndarray,
-               vbasis: BasisTuple,
-               w: FormParameters,
-               dx: ndarray) -> None:
-        for i in ix:
-            b[i] = np.sum(self.form(*vbasis[i], w) * dx, axis=1)
-
     def assemble(self,
-                 ubasis: Basis,
-                 vbasis: Optional[Basis] = None,
-                 w: Optional[Any] = (None, None, None),
-                 nthreads: Optional[int] = 1) -> ndarray:
+                 u: Basis,
+                 v: Optional[Basis] = None,
+                 w: Dict[str, DiscreteField] = {}) -> ndarray:
 
-        import threading
+        assert v is None
+        v = u
 
-        assert vbasis is None
-        vbasis = ubasis
+        nt = v.nelems
+        dx = v.dx
+        w = FormDict({**v.default_parameters(), **self.dictify(w)})
 
-        nt = vbasis.nelems
-        dx = vbasis.dx
-        w = self.parameters(w, vbasis)
+        # initialize COO data structures
+        sz = v.Nbfun * nt
+        data = np.zeros(sz)
+        rows = np.zeros(sz)
+        cols = np.zeros(sz)
 
-        data = np.zeros((vbasis.Nbfun, nt))
-        rows = np.zeros(vbasis.Nbfun * nt)
-        cols = np.zeros(vbasis.Nbfun * nt)
-
-        for i in range(vbasis.Nbfun):
-            # find correct location in data,rows,cols
+        for i in range(v.Nbfun):
             ixs = slice(nt * i, nt * (i + 1))
-            rows[ixs] = vbasis.element_dofs[i, :]
+            rows[ixs] = v.element_dofs[i]
             cols[ixs] = np.zeros(nt)
+            data[ixs] = self._kernel(v.basis[i], w, dx)
 
-        indices = range(vbasis.Nbfun)
+        return self._assemble_numpy_vector(data, rows, cols, (v.N, 1))
 
-        threads = [threading.Thread(
-            target=self.kernel,
-            args=(data, ix, vbasis.basis, w, dx))
-                   for ix in np.array_split(indices, nthreads, axis=0)]
-
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
-
-        return coo_matrix((data.flatten('C'), (rows, cols)),
-                          shape=(vbasis.N, 1)).toarray().T[0]
+    def _kernel(self, v, w, dx):
+        return np.sum(self.form(v, w) * dx, axis=1)
 
 
 def linear_form(form: Callable) -> LinearForm:
-    return LinearForm(form)
+
+    # for backwards compatibility
+    from .form_parameters import FormParameters
+
+    class ClassicLinearForm(LinearForm):
+
+        def _kernel(self, v, w, dx):
+            W = {k: w[k].f for k in w}
+            if 'w' in w:
+                W['dw'] = w['w'].df
+            if v.ddf is not None:
+                return np.sum(self.form(v=v.f, dv=v.df, ddv=v.ddf,
+                                        w=FormParameters(**W)) * dx, axis=1)
+            else:
+                return np.sum(self.form(v=v.f, dv=v.df,
+                                        w=FormParameters(**W)) * dx, axis=1)
+
+    return ClassicLinearForm(form)