+++ skfem/element/element_tri/element_tri_morley.py
@@ -16,10 +16,9 @@ class ElementTriMorley(ElementH2):
         elif i == 2:
             return u(*v[2])
         elif i == 3:
-            return du[0](*e[0])*n[0, 0] + du[1](*e[0])*n[0, 1]
+            return du[0](*e[0]) * n[0, 0] + du[1](*e[0]) * n[0, 1]
         elif i == 4:
-            return du[0](*e[1])*n[1, 0] + du[1](*e[1])*n[1, 1]
+            return du[0](*e[1]) * n[1, 0] + du[1](*e[1]) * n[1, 1]
         elif i == 5:
-            return du[0](*e[2])*n[2, 0] + du[1](*e[2])*n[2, 1]
-        else:
-            raise Exception("!")
+            return du[0](*e[2]) * n[2, 0] + du[1](*e[2]) * n[2, 1]
+        self._index_error()