+++ skfem/models/elasticity.py
@@ -1,23 +1,48 @@
-"""Linear elasticity."""
+"""Weak forms for elasticity."""
 
-from skfem.assembly import bilinear_form
+from skfem.assembly import BilinearForm
+from .helpers import ddot, trace, sym_grad,\
+    transpose, eye, grad
 
 
 def lame_parameters(E, nu):
-    return E / (2. * (1. + nu)), E * nu / ((1. + nu) * (1. - 2. * nu))
+    """Calculate the Lamé parameters from E and nu.
+
+    Parameters
+    ----------
+    E
+        Young's modulus
+    nu
+        Poisson ratio
+
+    Returns
+    -------
+    float
+        The first Lamé parameter
+    float
+        The second Lamé parameter
+
+    """
+    return (E / (2. * (1. + nu)),
+            E * nu / ((1. + nu) * (1. - 2. * nu)))
+
+
+def linear_stress(Lambda=1., Mu=1.):
+
+    def C(T):
+        """Linear-elastic stress-strain relationship."""
+        return 2. * Mu * T + Lambda * eye(trace(T), T.shape[0])
+
+    return C
 
 
 def linear_elasticity(Lambda=1., Mu=1.):
-    @bilinear_form
-    def weakform(u, du, v, dv, w):
-        from .helpers import ddot, trace, transpose, eye
-        
-        def C(T):
-            return 2.0 * Mu * T + Lambda * eye(trace(T), T.shape[0])
+    """Weak form of the linear elasticity operator."""
 
-        def Eps(dw):
-            return 0.5 * (dw + transpose(dw))
+    C = linear_stress(Lambda, Mu)
 
-        return ddot(C(Eps(du)), Eps(dv))
+    @BilinearForm
+    def weakform(u, v, w):
+        return ddot(C(sym_grad(u)), sym_grad(v))
 
     return weakform